Refactor:PlaceDetail 관련 폴더 및 파일 내용 리팩토링
