+++ app/feature/place/components/PlaceDetail/organisms/RelatedCurationSlider.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import SliderLayout from "@/common/components/layout/Slider/SliderLayout";
+import Title from "@/common/components/ui/text/Title";
 import { ReactNode } from "react";
 
 interface RelatedCurationSliderProps {
@@ -15,7 +16,7 @@ export default function RelatedCurationSlider({
 }: RelatedCurationSliderProps) {
   return (
     <div>
-      <span className="text-black headline2">{title}</span>
+      <Title title={title} className="headline2" />
       <SliderLayout className="mt-[1.6rem] mb-[6rem]">{children}</SliderLayout>
     </div>
   );
