+++ app/feature/place/components/PlaceDetail/organisms/RelatedPlaceSlider.tsx
@@ -1,4 +1,5 @@
 import SliderLayout from "@/common/components/layout/Slider/SliderLayout";
+import Title from "@/common/components/ui/text/Title";
 import { ReactNode } from "react";
 
 interface PlaceRelatedSliderProps {
@@ -13,7 +14,7 @@ export default function RelatedPlaceSlider({
 }: PlaceRelatedSliderProps) {
   return (
     <div>
-      <span className="text-black headline2">{title}</span>
+      <Title title={title} className="headline2" />
       <SliderLayout className="mt-[1.6rem] mb-[6rem]">{children}</SliderLayout>
     </div>
   );
